requirements.txt
# """
#Requirements file for Green Cover Calculator
#"""

# Core dependencies
numpy>=1.20.0              # Numerical computing
matplotlib>=3.3.0          # Visualization and plotting
scipy>=1.11                # KDE in NDVI histogram, spatial CV filter

# Optional dependencies
rasterio>=1.2.0           # GeoTIFF file handling
scipy>=1.6.0              # Scientific computing utilities
Pillow>=8.0.0             # Image processing
scikit-image>=0.18.0      # Image processing algorithms

# Development dependencies
pytest>=6.0.0             # Testing framework
pytest-cov>=2.12.0        # Coverage reports
black>=21.0               # Code formatting
flake8>=3.9.0             # Code linting


# Green Cover Calculator — Python dependencies
# Install all with:  pip install -r requirements.txt


# ── Config file parsing ────────────────────────────────────────
pyyaml>=6.0          # reads config.yaml

# ── Satellite data I/O ─────────────────────────────────────────
rasterio>=1.3        # reads GeoTIFF (Landsat) and JP2 (Sentinel-2)
                     # On Windows, install via conda:
                     #   conda install -c conda-forge rasterio

# ── Optional: faster GeoTIFF loading, reprojection ────────────
# gdal>=3.6           # usually installed as a dependency of rasterio

# ── Optional: Pillow fallback (TIF only, no JP2) ──────────────
Pillow>=10.0        # used only when rasterio is not available

